requirements.txt
streamlit
requests
pypdf
sentence-transformers
faiss-cpu
numpy